Create NewsApi and refactor NewsRepository to use it
