+++ lib/presentation/widgets/news_widgets/interesting_news.dart
@@ -1,13 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:tennis_cup/data/data_providers/news_api.dart';
 import 'package:tennis_cup/data/repositories/news_repository.dart';
 import 'package:tennis_cup/presentation/widgets/news_widgets/single_interesting_news.dart';
 
 class InterestingNews extends StatelessWidget {
+  final newsRepoSitory = const NewsRepository(newsApi: NewsApi());
   const InterestingNews({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final interestingNews = NewsRepository.fetchInterestingNews();
+    final interestingNews = newsRepoSitory.fetchInterestingNews();
+
     final width = MediaQuery.of(context).size.width;
     final height = MediaQuery.of(context).size.height;
     final isScreenHigh = height / width > 16 / 9;
